Refactor tests so python and java are separate
