+++ tests/test_validation_java.py
@@ -3,8 +3,6 @@ import os
 from pathlib import Path
 
 import pytest
-from jsonschema import validate
-from jsonschema.exceptions import ValidationError
 
 ROOT = Path(__file__).parent.parent
 os.environ["CLASSPATH"] = ":".join(
@@ -33,38 +31,28 @@ def runif_java_configured(func):
     )
 
 
-def test_validation_pass_python(schemas, qualifier, passing_example):
-    assert qualifier in schemas, f"{qualifier} missing from schemas"
-    validate(passing_example, schemas[qualifier])
-    # TODO: raise all validation errors for debugging, using IValidator
-    # interface. This requires knowing the JSON Schema spec ahead of time, for
-    # example by ensuring $schema is set.
-
-
-def test_validation_fail_python(schemas, qualifier, failing_example):
-    assert qualifier in schemas, f"{qualifier} missing from schemas"
-    with pytest.raises(ValidationError):
-        validate(failing_example, schemas[qualifier])
-
-
 @runif_java_configured
 def test_validation_pass_java(schemas, qualifier, passing_example):
+    assert qualifier in schemas, f"{qualifier} missing from schemas"
+
     JSONObject = autoclass("org.json.JSONObject")
     SchemaLoader = autoclass("org.everit.json.schema.loader.SchemaLoader")
 
     raw_schema = JSONObject(json.dumps(schemas[qualifier]))
-    example = JSONObject(json.dumps(passing_example))
     schema = SchemaLoader.load(raw_schema)
+    example = JSONObject(json.dumps(passing_example))
     schema.validate(example)
 
 
 @runif_java_configured
 def test_validation_fail_java(schemas, qualifier, failing_example):
+    assert qualifier in schemas, f"{qualifier} missing from schemas"
+
     JSONObject = autoclass("org.json.JSONObject")
     SchemaLoader = autoclass("org.everit.json.schema.loader.SchemaLoader")
 
     raw_schema = JSONObject(json.dumps(schemas[qualifier]))
-    example = JSONObject(json.dumps(failing_example))
     schema = SchemaLoader.load(raw_schema)
+    example = JSONObject(json.dumps(failing_example))
     with pytest.raises(JavaException):
         schema.validate(example)
